Missed an important use case.
Any time you add an ingredient to a recipe, we make a copy of that
ingredient. The actual code creates a new object using the existing sql
record. The changes I had made previously made a unpleasant mess of
this.

I added an optional parameter to the NamedEntity constructor called key.
If it is provided, we use that as the _key of the new object. If it
isn't, we use the key from the sql record. It isn't quite as pretty as I
had wanted, but it is still better than what we had.

I also cleaned up two deprecation warnings for qt 5.15 and a possible
dangling else warning.